Fix regression that exported isTracerSupportedByOS to a separate repo
