+++ pkg/ebpf/common_test.go
@@ -6,23 +6,6 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestVerifyKernelFuncs(t *testing.T) {
-	missing, err := verifyKernelFuncs("./testdata/kallsyms.supported")
-	assert.Empty(t, missing)
-	assert.Empty(t, err)
-
-	missing, err = verifyKernelFuncs("./testdata/kallsyms.unsupported")
-	assert.NotEmpty(t, missing)
-	assert.Empty(t, err)
-
-	missing, err = verifyKernelFuncs("./testdata/kallsyms.empty")
-	assert.NotEmpty(t, missing)
-	assert.Empty(t, err)
-
-	_, err = verifyKernelFuncs("./testdata/kallsyms.d_o_n_o_t_e_x_i_s_t")
-	assert.NotEmpty(t, err)
-}
-
 func TestSnakeToCamel(t *testing.T) {
 	for test, exp := range map[string]string{
 		"closed_conn_dropped":      "ClosedConnDropped",
